-fixed all remarks from comments -added re-export to all components -divide ProgressBar component to ProgressBar'; & CollectionProgress -changed font-size from .scss to _mixins font
