+++ src/modules/mainPageCollection/components/ActiveCollectionsCard/ActiveCollectionsCard.module.scss
@@ -42,8 +42,7 @@
   align-items: center;
   gap: 16px;
   font-weight: 600;
-  font-size: 16px;
-  line-height: 1.5;
+  @include font(16, 24);
 
   &::before {
     content: '';
@@ -77,8 +76,7 @@
 
 .title {
   font-weight: 700;
-  font-size: 20px;
-  line-height: 1.4;
+  @include font(20, 28);
   font-family: $secondary-font-family;
   padding-top: 16px;
 
@@ -87,16 +85,14 @@
   }
 
   @include media(desktop) {
-    font-size: 24px;
-    line-height: 1.16667;
+    @include font(24, 28);
   }
 }
 
 .desc,
 .peopleDonate {
   font-weight: 400;
-  font-size: 16px;
-  line-height: 1.5;
+  @include font(16, 24);
   display: flex;
   align-items: center;
   gap: 8px;
