+++ src/modules/mainPageCollection/index.js
@@ -1,1 +1,5 @@
 export { default as MainPageCollection } from './components/MainPageCollection/MainPageCollection';
+export { default as ActiveCollectionsCard } from './components/ActiveCollectionsCard/ActiveCollectionsCard';
+export { default as ActiveCollectionList } from './components/ActiveCollectionList/ActiveCollectionList';
+export { default as CollectionProgress } from './components/CollectionProgress/CollectionProgress';
+export { default as collections } from './data/collections.json';
