+++ src/shared/components/ProgressBar/ProgressBar.module.scss
@@ -1,76 +1,12 @@
-@import '../../styles/mixins';
-
-.mainProgressContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.targetContainer {
-  display: flex;
-  justify-content: space-between;
-  gap: 8px;
-
-  .targetText {
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 1.5;
-    display: flex;
-    flex-direction: column;
-
-    .counts {
-      font-weight: 700;
-      font-size: 20px;
-      font-family: $secondary-font-family;
-      line-height: 1.4;
-    }
-  }
-}
-
-.progressContainer {
-  display: flex;
-  width: 100%;
-  gap: 16px;
-  align-items: center;
-
-  @include media(mobile) {
-    gap: 24px;
-  }
-}
-
 .progressBarContainer {
-  max-width: 282px;
-  min-width: 210px;
-  height: 12px;
-  background-color: #e0e0e0;
   border-radius: 32px;
+  background-color: #e0e0e0;
   overflow: hidden;
-
-  @include media(mobile) {
-    min-width: 282px;
-  }
-
-  @include media(tablet) {
-    min-width: 236px;
-    max-width: 236px;
-  }
-
-  @include media(desktop) {
-    min-width: 320px;
-
-    max-width: 320px;
-  }
 }
 
 .progressBar {
   height: 100%;
-  background-color: black;
+  background-color: $swiper-active-dots-color;
   border-radius: 32px;
   transition: width 0.3s ease-in-out;
-}
-
-.rate {
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 1.2;
 }
